Remove promises and extract to react-dev-utils
