+++ packages/react-scripts/package.json
@@ -66,5 +66,6 @@
   },
   "optionalDependencies": {
     "fsevents": "1.0.17"
-  }
+  },
+  "proxy": "http://localhost:5000"
 }
