+++ packages/react-scripts/scripts/start.js
@@ -35,7 +35,7 @@ const paths = require('../config/paths');
 const config = require('../config/webpack.config.dev');
 const devServerConfig = require('../config/webpackDevServer.config');
 const createWebpackCompiler = require('./utils/createWebpackCompiler');
-const prepareProxy = require('./utils/prepareProxy');
+const prepareProxy = require('react-dev-utils/prepareProxy');
 
 const useYarn = fs.existsSync(paths.yarnLockFile);
 const cli = useYarn ? 'yarn' : 'npm';
@@ -73,24 +73,27 @@ function run(port) {
     }
   );
 
+  // Load proxy config
+  const proxy = require(paths.appPackageJson).proxy;
   // Serve webpack assets generated by the compiler over a web sever.
-  prepareProxy().then(proxy => {
-    const devServer = new WebpackDevServer(compiler, devServerConfig(proxy));
+  const devServer = new WebpackDevServer(
+    compiler,
+    devServerConfig(prepareProxy(proxy))
+  );
 
-    // Launch WebpackDevServer.
-    devServer.listen(port, host, err => {
-      if (err) {
-        return console.log(err);
-      }
+  // Launch WebpackDevServer.
+  devServer.listen(port, host, err => {
+    if (err) {
+      return console.log(err);
+    }
 
-      if (isInteractive) {
-        clearConsole();
-      }
-      console.log(chalk.cyan('Starting the development server...'));
-      console.log();
+    if (isInteractive) {
+      clearConsole();
+    }
+    console.log(chalk.cyan('Starting the development server...'));
+    console.log();
 
-      openBrowser(`${protocol}://${host}:${port}/`);
-    });
+    openBrowser(`${protocol}://${host}:${port}/`);
   });
 }
